+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [unreleased]
+### Added
+- Add option to start weeks in the date picker on Monday/Sunday. [#125]
 ### Changed
 - Greatly improved the context menu and handling.